pyproject.toml
[project]
name = "smolrl"
version = "0.1.0"
description = "A minimal, correct LLM-RL training library."
requires-python = ">=3.12, <3.14"
dependencies = [
    "torch>=2.4",
    "transformers>=4.45",
    "datasets>=3.0",
    "vllm>=0.6",
    "wandb>=0.18",
    "pyyaml>=6.0",
    "accelerate>=1.12.0",
    "math-verify[antlr4_13_2]",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.7",
]

[project.scripts]
tinyrl-train = "tinyrl.cli:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]